Fix the issue where, when using dynverse models, docker needed to use a tmp dir that was not available.
Fix the issue where, when using dynverse models, docker needed to use a tmp dir that was not available. The new command line adds a connection to the tmp dir.

This solution was only tested on MAC OS and Linux. The fix for Windows is going the be tested soon.