Updating the analysis of a single sample to allow reading and RDS file
Work still in progress. When finished, it should allow reading a dataset already clustered to explore markers and visualize the expression profiles.

I also removed the conversion of the sparse matrix to a matrix. This change was made as an attempt to reduce the memory consumption in the step that generated the heatmaps.